Reshuffle functions to better called modules + wip convert interactive from number to boolean
