+++ app/routes/jobs.$id.tools.reclustfcc.$module.tsx
@@ -1,4 +1,3 @@
-import { ClusterTable } from "~/tools/reclust";
 import type { LoaderArgs } from "@remix-run/node";
 import { json } from "@remix-run/node";
 import {
@@ -24,18 +23,19 @@ import {
   getParams,
   reclustfcc,
 } from "~/tools/reclustfcc.server";
-import { getPreviousCaprievalModule, moduleInfo } from "~/tools/shared";
 import { CompletedJobs } from "~/utils";
-import type { CaprievalPlotlyProps } from "~/tools/rescore.server";
-import {
-  getCaprievalPlots,
-  getPlotSelection,
-  getScores,
-} from "~/tools/rescore.server";
 import { ClientOnly } from "~/components/ClientOnly";
 import { CaprievalReport } from "~/components/Haddock3/CaprievalReport.client";
 import { ReWarning } from "~/components/ReWarning";
 import { getModuleDescriptions } from "~/catalogs/descriptionsFromSchema";
+import type { CaprievalPlotlyProps } from "~/models/caprieval.server";
+import {
+  getScores,
+  getPlotSelection,
+  getCaprievalPlots,
+} from "~/models/caprieval.server";
+import { moduleInfo } from "~/models/module_utils";
+import { ReClusterTable } from "~/tools/ReClusterTable";
 
 const fieldDescriptions = getModuleDescriptions(`clustfcc`, [
   "clust_cutoff",
@@ -72,30 +72,29 @@ export const loader = async ({ params, request }: LoaderArgs) => {
     token,
     moduleIndexPadding
   );
-  const caprievalModuleIndex = getPreviousCaprievalModule(
-    outputFiles,
-    moduleIndex,
-    interactivness
-  );
-  const scores = await getScores(
-    jobId,
-    interactivness ? moduleIndex : caprievalModuleIndex,
-    interactivness,
-    token,
-    moduleIndexPadding,
-    interactivness ? "clustfcc" : "caprieval"
-  );
-  const { scatterSelection, boxSelection } = getPlotSelection(request.url);
-  const plotlyPlots = await getCaprievalPlots(
-    jobId,
-    interactivness ? moduleIndex : caprievalModuleIndex,
-    interactivness,
-    token,
-    moduleIndexPadding,
-    scatterSelection,
-    boxSelection,
-    interactivness ? "clustrmsd" : "caprieval"
-  );
+  let scores;
+  let plotlyPlots;
+  if (interactivness > 0) {
+    scores = await getScores(
+      jobId,
+      moduleIndex,
+      interactivness,
+      token,
+      moduleIndexPadding,
+      "clustfcc"
+    );
+    const { scatterSelection, boxSelection } = getPlotSelection(request.url);
+    plotlyPlots = await getCaprievalPlots(
+      jobId,
+      moduleIndex,
+      interactivness,
+      token,
+      moduleIndexPadding,
+      scatterSelection,
+      boxSelection,
+      "clustrmsd"
+    );
+  }
 
   return json({
     moduleIndex,
@@ -148,7 +147,7 @@ export default function ReclusterPage() {
   } = useLoaderData<typeof loader>();
   const actionData = useActionData<typeof action>();
   // Strip SerializeObject<UndefinedToOptional wrapper
-  const plotlyPlotsStripped = plotlyPlots as CaprievalPlotlyProps;
+  const plotlyPlotsStripped = plotlyPlots as CaprievalPlotlyProps | undefined;
   const { state } = useNavigation();
   return (
     <>
@@ -227,9 +226,9 @@ export default function ReclusterPage() {
       <div>
         <details open={true}>
           <summary>Clusters</summary>
-          <ClusterTable clusters={clusters} />
+          <ReClusterTable clusters={clusters} />
         </details>
-        {scores && (
+        {scores && plotlyPlotsStripped && (
           <details open={true}>
             <summary>Capri evaluation</summary>
             <ClientOnly fallback={<p>Loading...</p>}>
